Add file existence check before downloading

+++ optuna/artifacts/_download.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import os
 import shutil
 
 from optuna._experimental import experimental_func
@@ -18,5 +19,8 @@ def download_artifact(artifact_store: ArtifactStore, artifact_id: str, file_path
         file_path:
             A path to save the downloaded artifact.
     """
+    if os.path.exists(file_path):
+        raise FileExistsError(f"File already exists: {file_path}")
+
     with artifact_store.open_reader(artifact_id) as reader, open(file_path, "wb") as writer:
         shutil.copyfileobj(reader, writer)
